refactor(clawdstrike-logos): extract report from verifier.rs
Move the verification report type model (VerificationBackend,
AttestationLevel, CheckOutcome, Conflict, ConsistencyResult,
CompletenessResult, WeakenedProhibition, InheritanceResult,
VerificationReport, DEFAULT_EXPECTED_ACTION_TYPES) and the three free
report builders (compute_attestation_level, build_policy_report,
report_backend) into verifier/report.rs. Pure relocation: the public
types are re-exported explicitly from the verifier module root and the
private builders are widened to pub(crate) so sibling modules reach them
via the root's pub(crate) use glob, mirroring the api_server precedent.
